NMSC-8 - Move services
