+++ src/app/landing/landing.component.ts
@@ -1,10 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import {Component} from '@angular/core';
 
 @Component({
   selector: 'app-landing',
   templateUrl: './landing.component.html',
   styleUrls: ['./landing.component.scss']
 })
-export class LandingComponent{
+export class LandingComponent {
   categories = [ 'ingredients', 'recipes', 'tools' ];
 }
