+++ src/app/services/index.ts
@@ -1,0 +1,2 @@
+export * from './ingredients.service';
+export * from './recipes.service';
